# MAINT: use the fast elementwise 2-norm implementation more often in linalg.norm