This applies the speedup in https://github.com/numpy/numpy/pull/3895 across more combinations of inputs, addressing the 'lovely example' in https://github.com/scipy/scipy/pull/4642#issuecomment-83237655.

Some timings:
```
before:
In [1]: import numpy as np
In [2]: a = np.random.randn(1000, 1000)
In [3]: %timeit np.linalg.norm(a.ravel())
1000 loops, best of 3: 441 µs per loop
In [4]: %timeit np.linalg.norm(a, 'fro')
1000 loops, best of 3: 1.84 ms per loop

after:
In [1]: import numpy as np
In [2]: a = np.random.randn(1000, 1000)
In [3]: %timeit np.linalg.norm(a.ravel())
1000 loops, best of 3: 441 µs per loop
In [4]: %timeit np.linalg.norm(a, 'fro')
1000 loops, best of 3: 441 µs per loop
```